requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
oauth==1.0.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
